+++ main/world_map/world_map_helpers.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//01/04/26
+//04/04/26
 
 /* exported selPoint, tooltipPoint, tooltipPanel, selFindPoint, tooltipFindPoint, biographyCheck, saveLibraryTags, wheelResize, headerCountryName, headerCoords, drawHeader, drawTaggingPoint, paintLayers */
 
@@ -542,6 +542,18 @@ const imgAsync = {
 	lowMemMode: { maxSize: 1000 },
 	fullImg: null
 };
+/**
+ * Paint country layers
+ *
+ * @function
+ * @name fillSubLayer
+ * @kind variable
+ * @param {GdiBitmap} subLayer
+ * @param {string} id
+ * @param {string} mode
+ * @param {number} scale
+ * @returns {void}
+ */
 const fillSubLayer = (subLayer, id, mode, scale = Math.min(imgAsync.layers.w / worldMap.imageMap.Width, imgAsync.layers.h / worldMap.imageMap.Height)) => {
 	if (!mode || !mode.length) { return; }
 	const flagSize = 64; const w = 40; const h = 30;
@@ -578,7 +590,8 @@ const fillSubLayer = (subLayer, id, mode, scale = Math.min(imgAsync.layers.w / w
 		}
 		case 'flag': {
 			const w = imgAsync.layers.w / 2; const h = imgAsync.layers.h / 2;
-			const x = point.x * scale - w / 2; const y = point.y * scale - h / 2;
+			const x = point.xCorr * scale - w / 2; const y = point.yCorr * scale - h / 2;
+			layerGr.SetInterpolationMode(InterpolationMode.NearestNeighbor);
 			layerGr.DrawImage(flag, x, y, w, h, 0, 0, flag.Width, flag.Height);
 			break;
 		}